UIBULKED-38: refactor contants
