+++ src/constants/constants.js
@@ -210,6 +210,7 @@ export const ACTIONS = {
 export const OPTIONS = {
   TEMPORARY_LOCATION: 'TEMPORARY_LOCATION',
   PERMANENT_LOCATION: 'PERMANENT_LOCATION',
+  STATUS: 'STATUS',
 };
 
 export const ITEMS_OPTIONS = [
@@ -225,60 +226,7 @@ export const ITEMS_OPTIONS = [
   },
   {
     value: 'STATUS',
-    label: 'ui-bulk-edit.layer.options.itemStatus',
-    disabled: false,
-  },
-];
-
-export const ITEM_STATUS_OPTIONS = [
-  {
-    value: '',
-    label: 'ui-bulk-edit.layer.options.placeholder',
-    disabled: true,
-  },
-  {
-    value: 'Available',
-    label: 'ui-bulk-edit.layer.options.available',
-    disabled: false,
-  },
-  {
-    value: 'Missing',
-    label: 'ui-bulk-edit.layer.options.missing',
-    disabled: false,
-  },
-  {
-    value: 'Withdrawn',
-    label: 'ui-bulk-edit.layer.options.withdrawn',
-    disabled: false,
-  },
-  {
-    value: 'In process (non-requestable)',
-    label: 'ui-bulk-edit.layer.options.non-requestable',
-    disabled: false,
-  },
-  {
-    value: 'Intellectual item"',
-    label: 'ui-bulk-edit.layer.options.intellectual',
-    disabled: false,
-  },
-  {
-    value: 'Long missing',
-    label: 'ui-bulk-edit.layer.options.longMissing',
-    disabled: false,
-  },
-  {
-    value: 'Restricted',
-    label: 'ui-bulk-edit.layer.options.restricted',
-    disabled: false,
-  },
-  {
-    value: 'Unavailable',
-    label: 'ui-bulk-edit.layer.options.unavailable',
-    disabled: false,
-  },
-  {
-    value: 'Unknown',
-    label: 'ui-bulk-edit.layer.options.unknown',
+    label: 'ui-bulk-edit.layer.options.statusLabel',
     disabled: false,
   },
 ];
