+++ src/constants/index.js
@@ -1,2 +1,3 @@
 export * from './optionsRecordIdentifiers';
 export * from './constants';
+export * from './itemStatusOptions';
