+++ src/constants/itemStatusOptions.js
@@ -1,0 +1,52 @@
+export const ITEM_STATUS_OPTIONS = [
+  {
+    value: '',
+    label: 'ui-bulk-edit.layer.options.placeholder',
+    disabled: true,
+  },
+  {
+    value: 'Available',
+    label: 'ui-bulk-edit.layer.options.available',
+    disabled: false,
+  },
+  {
+    value: 'Missing',
+    label: 'ui-bulk-edit.layer.options.missing',
+    disabled: false,
+  },
+  {
+    value: 'Withdrawn',
+    label: 'ui-bulk-edit.layer.options.withdrawn',
+    disabled: false,
+  },
+  {
+    value: 'In process (non-requestable)',
+    label: 'ui-bulk-edit.layer.options.non-requestable',
+    disabled: false,
+  },
+  {
+    value: 'Intellectual item"',
+    label: 'ui-bulk-edit.layer.options.intellectual',
+    disabled: false,
+  },
+  {
+    value: 'Long missing',
+    label: 'ui-bulk-edit.layer.options.longMissing',
+    disabled: false,
+  },
+  {
+    value: 'Restricted',
+    label: 'ui-bulk-edit.layer.options.restricted',
+    disabled: false,
+  },
+  {
+    value: 'Unavailable',
+    label: 'ui-bulk-edit.layer.options.unavailable',
+    disabled: false,
+  },
+  {
+    value: 'Unknown',
+    label: 'ui-bulk-edit.layer.options.unknown',
+    disabled: false,
+  },
+];
